Package by feature
Make directories according to feature and have relevant files in a
directory

Move hooks to new file
Seperate hooks from services, so they can be imployed by other services
if required
