+++ src/iot-device/constants.ts
