+++ src/iot-device/iot-device-model.ts
