+++ src/user/user.ts
@@ -1,0 +1,5 @@
+export interface IUser {
+  username: string;
+  password: string;
+  group: string;
+}
